backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
python-dotenv
pytest
httpx
